# docs/design/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Team Document Internal API
  version: 1.0.0
  description: Browser-facing internal contract. Not a customer Public API.
servers:
  - url: /api/v1
security:
  - sessionCookie: []
...
paths:
...
  /workspaces/{workspaceId}/documents:
    post:
      tags: [Document]
      operationId: createDocument
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: {type: string, minLength: 1, maxLength: 500}
                parentId: {$ref: '#/components/schemas/NullableId'}
                afterDocumentId: {$ref: '#/components/schemas/NullableId'}
      responses:
        '201': {description: Created, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/tree:
    get:
      tags: [Document]
      operationId: getDocumentTree
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
      responses:
        '200': {description: Permission-filtered tree, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentTree'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/trash:
    get:
      tags: [Document]
      operationId: listTrashedDocuments
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {description: Trashed documents, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentPage'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}:
    get:
      tags: [Document]
      operationId: getDocument
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/Cursor'}]
      responses:
        '200': {description: OK, content: {application/json: {schema: {$ref: '#/components/schemas/DocumentDetail'}}}}
        default: {$ref: '#/components/responses/Problem'}
    put:
      tags: [Document]
      operationId: updateDocumentMetadata
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/UpdateDocumentMetadata'}
      responses:
        '200': {description: Updated, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Document]
      operationId: purgeDocument
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/Reason'}
      responses:
        '202': {description: Purge accepted, content: {application/json: {schema: {$ref: '#/components/schemas/JobReference'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/trash:
    post:
      tags: [Document]
      operationId: trashDocument
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/Reason'}
      responses:
        '200': {description: Trashed, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/restore:
    post:
      tags: [Document]
      operationId: restoreDocument
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/RestoreDocument'}
      responses:
        '200': {description: Restored, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/move-preview:
    post:
      tags: [Document]
      operationId: previewDocumentMove
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody: {$ref: '#/components/requestBodies/MoveDocument'}
      responses:
        '200': {description: Impact, content: {application/json: {schema: {$ref: '#/components/schemas/ImpactPreview'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/move:
    post:
      tags: [Document]
      operationId: moveDocument
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/MoveDocumentInput'}, {type: object, required: [previewToken], properties: {previewToken: {type: string}}}]}}}
      responses:
        '200': {description: Moved, content: {application/json: {schema: {$ref: '#/components/schemas/Document'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /workspaces/{workspaceId}/documents/{documentId}/draft:
    get:
      tags: [Document]
      operationId: getDraft
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}]
      responses:
        '200': {description: Draft, content: {application/json: {schema: {$ref: '#/components/schemas/Draft'}}}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Document]
      operationId: createOrGetDraft
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      responses:
        '200': {description: Draft, content: {application/json: {schema: {$ref: '#/components/schemas/Draft'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/lease:
    post:
      tags: [Document]
      operationId: acquireEditLease
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [clientInstanceId], properties: {clientInstanceId: {$ref: '#/components/schemas/Id'}, force: {type: boolean, default: false}, reason: {type: string, maxLength: 500}}}}}
      responses:
        '200': {description: Lease, content: {application/json: {schema: {$ref: '#/components/schemas/EditLease'}}}}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Document]
      operationId: releaseEditLease
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/LeaseToken'}, {$ref: '#/components/parameters/ClientInstance'}, {$ref: '#/components/parameters/CsrfToken'}]
      responses:
        '204': {description: Released}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/lease/renew:
    post:
      tags: [Document]
      operationId: renewEditLease
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/LeaseToken'}, {$ref: '#/components/parameters/ClientInstance'}, {$ref: '#/components/parameters/CsrfToken'}]
      responses:
        '200': {description: Renewed, content: {application/json: {schema: {$ref: '#/components/schemas/EditLease'}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspaceId}/documents/{documentId}/draft/operations:
    post:
      tags: [Document]
      operationId: applyDraftOperations
      parameters: [{$ref: '#/components/parameters/WorkspaceId'}, {$ref: '#/components/parameters/DocumentId'}, {$ref: '#/components/parameters/IfMatch'}, {$ref: '#/components/parameters/IdempotencyKey'}, {$ref: '#/components/parameters/LeaseToken'}, {$ref: '#/components/parameters/ClientInstance'}, {$ref: '#/components/parameters/CsrfToken'}]
      requestBody:
        required: true
        content: {application/json: {schema: {type: object, required: [operations], properties: {operations: {type: array, minItems: 1, items: {$ref: '#/components/schemas/DocumentOperation'}}}}}}
      responses:
        '200': {description: Applied, content: {application/json: {schema: {$ref: '#/components/schemas/MutationResult'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
components:
  securitySchemes:
    sessionCookie: {type: apiKey, in: cookie, name: adoc_session}
  parameters:
    WorkspaceId: {name: workspaceId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
    DocumentId: {name: documentId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
...
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 2048}}
...
    IfMatch: {name: If-Match, in: header, required: true, schema: {type: string, pattern: '^"[0-9]+"$'}}
    IdempotencyKey: {name: Idempotency-Key, in: header, required: true, schema: {type: string, minLength: 16, maxLength: 128}}
    CsrfToken: {name: X-CSRF-Token, in: header, required: true, schema: {type: string, minLength: 43, maxLength: 1024}}
    LeaseToken: {name: X-Edit-Lease, in: header, required: true, schema: {type: string}}
    ClientInstance: {name: X-Client-Instance, in: header, required: true, schema: {$ref: '#/components/schemas/Id'}}
  requestBodies:
...
    MoveDocument:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/MoveDocumentInput'}}}
...
    Reason:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [reason], properties: {reason: {type: string, minLength: 1, maxLength: 500}}}}}
...
    UpdateDocumentMetadata:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [title], properties: {title: {type: string, minLength: 1, maxLength: 500}}}}}
    RestoreDocument:
      required: true
      content: {application/json: {schema: {type: object, additionalProperties: false, required: [parentId], properties: {parentId: {$ref: '#/components/schemas/NullableId'}, afterDocumentId: {$ref: '#/components/schemas/NullableId'}}}}}
...
  responses:
    Problem:
      description: Domain or transport problem
      content: {application/problem+json: {schema: {$ref: '#/components/schemas/Problem'}}}
  schemas:
    Id: {type: string, format: uuid}
    NullableId: {oneOf: [{$ref: '#/components/schemas/Id'}, {type: 'null'}]}
...
    Document: {type: object, additionalProperties: false, required: [id, title, parentId, status, currentVersionId, revision], properties: {id: {$ref: '#/components/schemas/Id'}, title: {type: string}, parentId: {$ref: '#/components/schemas/NullableId'}, status: {type: string, enum: [ACTIVE, TRASHED, PURGING]}, currentVersionId: {$ref: '#/components/schemas/NullableId'}, revision: {type: integer, minimum: 0}}}
    DocumentPage: {type: object, additionalProperties: false, required: [items], properties: {items: {type: array, items: {$ref: '#/components/schemas/Document'}}, nextCursor: {type: [string, 'null']}}}
    DocumentTree: {type: object, additionalProperties: false, required: [nodes, watermark], properties: {nodes: {type: array, items: {$ref: '#/components/schemas/DocumentTreeNode'}}, watermark: {type: integer, minimum: 0}}}
    DocumentTreeNode: {type: object, additionalProperties: false, required: [document, children], properties: {document: {$ref: '#/components/schemas/Document'}, children: {type: array, items: {$ref: '#/components/schemas/DocumentTreeNode'}}}}
    JobReference: {type: object, additionalProperties: false, required: [jobId, status], properties: {jobId: {$ref: '#/components/schemas/Id'}, status: {type: string, enum: [QUEUED]}}}
    DocumentDetail: {allOf: [{$ref: '#/components/schemas/Document'}, {type: object, properties: {draft: {oneOf: [{$ref: '#/components/schemas/Draft'}, {type: 'null'}]}, publishedVersion: {oneOf: [{$ref: '#/components/schemas/PublishedVersion'}, {type: 'null'}]}}}]}
    Draft: {type: object, additionalProperties: false, required: [id, documentId, baseVersionId, revision, schemaVersion, contentFingerprint, content], properties: {id: {$ref: '#/components/schemas/Id'}, documentId: {$ref: '#/components/schemas/Id'}, baseVersionId: {$ref: '#/components/schemas/NullableId'}, revision: {type: integer, minimum: 0}, schemaVersion: {type: integer, minimum: 1}, contentFingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}, content: {$ref: '../contracts/document-content.schema.json'}}}
    EditLease: {type: object, additionalProperties: false, required: [holderUserId, clientInstanceId, expiresAt, revision], properties: {holderUserId: {$ref: '#/components/schemas/Id'}, clientInstanceId: {$ref: '#/components/schemas/Id'}, token: {type: string, writeOnly: true, minLength: 43, maxLength: 43}, expiresAt: {type: string, format: date-time}, revision: {type: integer, minimum: 0}}}
    DocumentOperation: {$ref: '../contracts/document-operation.schema.json'}
    MutationResult: {type: object, additionalProperties: false, required: [revision, contentFingerprint, appliedOperationIds, inverseOperations], properties: {revision: {type: integer, minimum: 1}, contentFingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}, appliedOperationIds: {type: array, items: {$ref: '#/components/schemas/Id'}}, inverseOperations: {type: array, items: {$ref: '#/components/schemas/DocumentOperation'}}}}
    PublishedVersion: {type: object, additionalProperties: false, required: [id, documentId, number, publishedAt, publisherId, schemaVersion, contentFingerprint, basedOnVersionId, sourceDraftRevision, content, summary, reviewSnapshot, discussionIds], properties: {id: {$ref: '#/components/schemas/Id'}, documentId: {$ref: '#/components/schemas/Id'}, number: {type: integer, minimum: 1}, publishedAt: {type: string, format: date-time}, publisherId: {$ref: '#/components/schemas/Id'}, schemaVersion: {type: integer, minimum: 1}, contentFingerprint: {type: string, pattern: '^[a-f0-9]{64}$'}, basedOnVersionId: {$ref: '#/components/schemas/NullableId'}, sourceDraftRevision: {type: integer, minimum: 0}, content: {$ref: '../contracts/document-content.schema.json'}, summary: {type: string, minLength: 1, maxLength: 1000}, reviewSnapshot: {type: object}, discussionIds: {type: array, items: {$ref: '#/components/schemas/Id'}}}}
...
    MoveDocumentInput: {type: object, additionalProperties: false, required: [newParentId, afterDocumentId], properties: {newParentId: {$ref: '#/components/schemas/NullableId'}, afterDocumentId: {$ref: '#/components/schemas/NullableId'}}}
    ImpactPreview: {type: object, additionalProperties: false, required: [previewToken, permissionChanges, policyChanges, expiresAt], properties: {previewToken: {type: string, minLength: 43, maxLength: 43}, permissionChanges: {type: integer, minimum: 0}, policyChanges: {type: integer, minimum: 0}, expiresAt: {type: string, format: date-time}}}
...
    Problem: {type: object, required: [type, title, status, code, retryable, correlationId], properties: {type: {type: string, format: uri}, title: {type: string}, status: {type: integer}, code: {type: string}, retryable: {type: boolean}, correlationId: {type: string}, currentRevision: {type: integer}, baseVersionId: {$ref: '#/components/schemas/NullableId'}, currentVersionId: {$ref: '#/components/schemas/NullableId'}, draftId: {$ref: '#/components/schemas/NullableId'}, fieldErrors: {type: array, items: {type: object, required: [field, code], properties: {field: {type: string}, code: {type: string}}}}}}